Add test to make sure all project lists agree.
Project lists are:
* .gitmodules
* tools/git-remotes
* tools/BUILD.bazel

Added a symlink to .gitmodules in tools/gitmodules. This is needed
because we don't have access to `//:.gitmodules`. We could easily have
access to it, but that means adding it to the toplevel BUILD.bazel, which
in turn means .gitmodules becomes part of the minimal bootstrap set we
need for building the Docker image. We'd like to avoid having to rebuild
the Docker image from scratch (6GB upload) whenever we add/remove/rename
a submodule.