perf: Attempt to reduce disk usage for building docker images.
There's about 661MiB of docs in the haskell distribution, which is
copied over twice, so it's 1322MiB of docs (out of currently 7.5GiB of
bazel cache after the initial aquery).

GitHub Actions runs out of disk space when producing the layers for
upload. Afterwards, during CI runs, it's fine, but while
uploading/compressing layers, it needs to replicate all the files in the
layer, so it takes up a lot of additional disk space all at once.

Another future thing we can do is split the layer, e.g. by doing `query`
and then `aquery` to first download some things and then extract them
and download more.